[SearchAPI]검증 스키마 리팩토링
